rumpcomp_pci_dmalloc creates unencrypted mappings
When SEV is enabled, the guest memory is encrypted and the
virtual devices (implemented by the hypervisors) cannot access
the memory.

As a work around, the kernel can use bounce buffer mechanisms
where the bounce buffer is unencrypted.

The bounce buffer and synchronization mechanism is implmented in
the patched version of NetBSD v8 (See https://github.com/shpark/netbsd-src).

In particular, the rumpkernel expects that memory regions
allocated by rumpcomp_pci_dmalloc be unencrypted, and it
uses those unencrypted regions as bounce buffers.

This patch introduces MapShared operation that are almost same
as Map operation, but leaves C-bit cleared in the PTE.

Status:
- Runs Redis in an SEV VM with e1000 NIC.
- CANNOT run Redis with virtio-net. Probably this is due to bugs in
  NetBSD virtio-net driver?